fix: AppManager 역할 분리
- 테마 및 뱃지,  레벨 계산 로직을 별도 helpers 모듈로 분리

close #144
