+++ src/stores/machines/service.ts
@@ -1,0 +1,7 @@
+import { interpret, type InterpreterFrom } from 'xstate';
+import { globalMachine } from './globalMachine';
+
+const service = interpret(globalMachine);
+
+export const getService = (): InterpreterFrom<typeof globalMachine>  => service;
+
